+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.14.7 (2019-10-03)
+* Added production render override option
+
+## v0.14.6 (2019-10-03)
+* Added emergency exception trace in production mode
+
 ## v0.14.5 (2019-10-03)
 * Fixed identifier lists